requirements.txt
pillow
python-casacore
monotonic
astropy
ephem
matplotlib
#aplpy
scipy